QEMU NVME: allow STag SMPS to go to maximum size
Changed 32 bit type to 64 bits to allow for the maximum 8TB page size.

Reported-by: Sarah Jelinek <sarah.jelinek@intel.com>

+++ hw/nvme.h
@@ -470,7 +470,7 @@ typedef struct NVMEAonPD {
 
 typedef struct NVMEAonStag {
     uint32_t pdid;
-    uint32_t smps;
+    uint64_t smps;
     uint64_t prp;
     uint64_t nmp;
 } NVMEAonStag;
